homework/fe-course/requirements.txt
numpy==1.26.4
python-dotenv==1.0.1
jupyter==1.0.0
